+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-03-xx
 
+## [0.55.0] - 2025-03-07
+
 ### Added
 
   - [ET-2141] `get_unsent_messages_ids_in_queue` method on `MailSession` to get MessageIds of unsent messages for given user_id